ARUHA-62: moved partitions endpoints to another URL; refactoring;
