+++ src/main/java/de/zalando/aruha/nakadi/controller/TopicsController.java
@@ -2,7 +2,6 @@ package de.zalando.aruha.nakadi.controller;
 
 import com.codahale.metrics.annotation.Timed;
 import de.zalando.aruha.nakadi.NakadiException;
-import de.zalando.aruha.nakadi.domain.Problem;
 import de.zalando.aruha.nakadi.repository.TopicRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -37,53 +36,6 @@ public class TopicsController {
         }
         catch (final NakadiException e) {
             return status(HttpStatus.SERVICE_UNAVAILABLE).body(e.getProblemMessage());
-        }
-    }
-
-    @Timed(name = "get_partitions", absolute = true)
-    @RequestMapping(value = "/{topic}/partitions", method = RequestMethod.GET)
-    public ResponseEntity<?> listPartitions(@PathVariable("topic") final String topic,
-                                            @RequestHeader(name = "X-Flow-Id", required = false) final String flowId) {
-        LOG.trace("Get partitions endpoint for topic '{}' is called for flow id: {}", topic, flowId);
-        try {
-            if (topicRepository.topicExists(topic)) {
-                return ok().body(topicRepository.listPartitions(topic));
-            }
-            else {
-                return status(HttpStatus.NOT_FOUND).body(new Problem("topic not found"));
-            }
-        }
-        catch (final NakadiException e) {
-            return status(HttpStatus.SERVICE_UNAVAILABLE).body(e.getProblemMessage());
-        }
-        catch (final Exception e) {
-            return status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-        }
-    }
-
-    @Timed(name = "get_partition", absolute = true)
-    @RequestMapping(value = "/{topic}/partitions/{partition}", method = RequestMethod.GET)
-    public ResponseEntity<?> getPartition(@PathVariable("topic") final String topic,
-                                          @PathVariable("partition") final String partition,
-                                          @RequestHeader(name = "X-Flow-Id", required = false) String flowId) {
-        LOG.trace("Get partition endpoint for topic '{}', partition '{}' is called for flow id: {}", topic, partition,
-                flowId);
-        try {
-            if (!topicRepository.topicExists(topic)) {
-                return status(HttpStatus.NOT_FOUND).body(new Problem("topic not found"));
-            }
-            else if (!topicRepository.partitionExists(topic, partition)) {
-                return status(HttpStatus.NOT_FOUND).body(new Problem("partition not found"));
-            }
-            else {
-                return ok().body(topicRepository.getPartition(topic, partition));
-            }
-        }
-        catch (final NakadiException e) {
-            return status(HttpStatus.SERVICE_UNAVAILABLE).body(e.getProblemMessage());
-        }
-        catch (final Exception e) {
-            return status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
     }
