+++ src/main/java/de/zalando/aruha/nakadi/repository/TopicRepository.java
@@ -23,17 +23,15 @@ public interface TopicRepository {
 
     boolean topicExists(String topic) throws NakadiException;
 
-    boolean areCursorsCorrect(String topic, List<Cursor> cursors);
+    boolean areCursorsCorrect(String topic, List<Cursor> cursors) throws NakadiException;
 
     boolean partitionExists(String topic, String partition) throws NakadiException;
 
     void postEvent(String topicId, String partitionId, String payload) throws NakadiException;
 
-    List<TopicPartition> listPartitions(String topicId);
+    List<TopicPartition> listPartitions(String topicId) throws NakadiException;
 
     TopicPartition getPartition(String topicId, String partition) throws NakadiException;
-
-    boolean validateOffset(String offsetToCheck, String newestOffset, String oldestOffset);
 
     EventConsumer createEventConsumer(String topic, Map<String, String> cursors);
 }
